backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy>=2.0.25
psycopg2-binary
pydantic
python-nmap==0.7.1
requests==2.31.0
python-owasp-zap-v2.4==0.0.20

python-dotenv==1.0.1
passlib==1.7.4
bcrypt==3.2.2
PyJWT==2.12.1
reportlab==4.1.0
slowapi==0.1.9
apscheduler==3.10.4
pyotp==2.9.0
cryptography==42.0.0
celery==5.3.6
redis==5.0.1
networkx>=3.3
scikit-learn>=1.4.0